requirements.txt
aiofiles==0.4.0
aiohttp==3.6.0
aiohttp-socks==0.2.2
aiorpcX==0.18.3
async-timeout==3.0.1
attrs==19.1.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.5
django-filter==2.2.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
dnspython==1.16.0
drf-yasg==1.16.1
ecdsa==0.13.2
fasteners==0.15
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
jsonrpclib-pelix==0.4.0
MarkupSafe==1.1.1
monotonic==1.5
multidict==4.5.2
node-semver==0.7.0
npmdownloader==1.1.1
openapi-codec==1.3.2
packaging==19.1
Pillow==6.1.0
protobuf==3.9.1
psycopg2==2.8.3
pyaes==1.6.1
PyJWT==1.7.1
pyparsing==2.4.2
pytz==2019.2
QDarkStyle==2.6.8
qrcode==6.1
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.1.2
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
yarl==1.3.0
